WIP: Update latest test build. Only small cosmetic changes... ... in this commit to d2r2 (no additional runtime changes). I tried some things, they are no good. There's little that can be done about <ComponentRouter/> performance I think. The main issue is appropriately implementing `componentShouldUpdate` React component lifecycle method:
https://reactjs.org/docs/react-component.html#shouldcomponentupdate

Do not be fooled into thinking the React.PureComponent will help
us much. It will not and cannot handle what we need in high density
tables where there are litterally thousands of mounted React
components. It is not feasible to efficiently re-render them
all on every event.

Good news is that it's rather simple to implement a general solution
that works great. However, you have to build it into your view data
model (i.e. the stuff passed through `this.props.renderData`).
Typically, we'll use a monotonically-increasing counter to
indicate that a namespace (data) has been written and then
keep track of this count in component instance state (i.e. this.count not this.state.count) and then answer `shouldComponentRender` w/a trivial numerical comparison.

This will eliminate all the re-render on most branches of very dense data trees for most operations and provide a huge and noticeable perf boost.