TEST RELEASE v0.0.46-crossroads buildID "J9LNSLDoR5ehnP9TZMM1kA".
Includes additional cleanup/refactoring ahead of planned continuation
of work on CellProcessManager's handling of CellProcessProxy helpers.
We must have the ability to embed a CellProcessProxy within an APM
as a helper cell. And, then be able to embed that APM (including its embedded CellProcessProxy helper) in yet another APM as a heler... And, so on. And, so on, and so on... More about this later this week; much of last week was devoted to shoring up test coverage to ensure the complex work required to address current restrictions on CellProcessProxy helper nesting / cell composition patterns doesn't go completely sideways.

Separately, this commit also contains some very small changes to the @encapsule/holistic-app-client-cm RTL that alter the way that the current revision of the HolisticAppClientKernel starts its child subprocesses when its process is started.

Now, HolisticAppClientKernel will automatically start DOMLocationProcessor, ClientDisplayAdaptor, and ClientViewProcessor (to be explained in more detail later) when it is started (currently done via external call to CellProcessor.act in `SOURCES/CLIENT/client.js`.

```
    boot1_start_kernel: {
      description: "Start core client app kernel subsystems.",
      actions: {
        enter: [{
          holarchy: {
            CellProcessor: {
              process: {
                create: {
                  apmID: "-1Ptaq_zTUa8Gfv_3ODtDg"
                  /* "Holistic App Client Kernel: DOM Location Processor" */

                }
              }
            }
          }
        }, {
          holarchy: {
            CellProcessor: {
              process: {
                create: {
                  apmID: "Hsu-43zBRgqHItCPWPiBng"
                  /* "Holistic App Client Kernel: Client View Processor" */

                }
              }
            }
          }
        }, {
          holarchy: {
            CellProcessor: {
              process: {
                create: {
                  apmID: "IxoJ83u0TXmG7PLUYBvsyg"
                  /* "Holistic Client App Kernel: d2r2/React Client Display Adaptor" */

                }
              }
            }
          }
        }]
      },
```