WIP: Updates to @encapsule/holarchy Cell Process Manager implementation to align the CellModel SiC programming model. Most of the changes should be transparent in terms of functionality. There are however many changes to CPM ControllerAction and TransitionOperator signatures in this build. And, the current calling conventions should not yet be considered stable as I am still making changes to these signatures and going through the tests (currently stable).
PLEASE DO NOT invest time porting existing code to this
build; I am publishing it only so you can take a look and
see the direction this is headed.

When I merge v0.0.47-alexandrite into the #release-test
branch then we can safely migrate existing CellModel's to
new CPM programming model and calling conventions. The intent
is to move everything to v0.0.47 and then use that to help
confirm changes in v0.0.48 that will effectively virtualize
CellProcessProxy.