TEST RELEASE v0.0.47-alexandrite buildID "jUyXyJapRpKXwtD0_4QRlA"
WIP on the following RTL's:

- @encapsule/holism
- @encapsule/d2r2
- @encapsule/d2r2-components

This work-in-progress is required to shore-up @encapsule/holism HTTP 1.1 Node.js server deployments that host (sort of not really at this point) the backend of a "holistic application".

Today, we wire up the backend in a derived app by-hand against the undocumented @encapsule/holism RTL (meaning mostly I do it). This is 100% BS and there's a plan for that. But, not enough time at the moment to make it much simpler for everyone.

Meanwhile, in the derived app there's a need to take 100% control of the d2r2/React rendering story against the DOM actual (via React wrapped in d2r2) and add another layer of abstraction on top of the low-level d2r2 routing protol to support integration with @encapule/holarchy CellProcessor cells.

And, because we pre-render via ReactDOM on the app server (and always will for a whole bunch of reasons unrelated to current client-side efforts) we need to tweak these low-level RTL's.

In the not-too-distant future most of the backend (including @encapsule/holism) will dissapear into @encapsule/holistic-app-server-cm RTL (just like all the reusable pieces of the HTML5 app client are getting sucked into @encapsule/holistic-app-client-cm RTL).