TEST-RELEASE v0.0.49-spectrolite builidID "Ipk3ibwPRriI9WKZn1_hIA"
This should be considered still WIP that is not quite ready for
use as intended. I am sharing it now because it's reached the
point where it's API complete to a first approximation.

When I am satisfied that all the little bugs are fixed and have
finished some cleanup, I will promote another build to
the release-test branch.

Tons of complicated work here uncluster a matrix of concerns
that span architecture, design, implementation, where you put
your files, how you think about even what you do with this stuff..