TEST RELEASE v0.0.49-spectrolite buildID "8_il5qWZQ1OLKFcBSrygmA"
Updates to @encapsule/holistic-nodejs-service RTL and its
embedded @encapsule/holism HTTP 1.1 request processor to
deal with the 1st of a set of 4-5 little critical integrations
that need to get patched up to get the matrix of backend
extension points defined by HolisticNodeService and HolisticCoreService
working correctly w/the newly re-integrated and recently-revised
@encapsule/holism RTL wrt:
- User identity assertion (i.e. the app-defined data types and synchronous function callback used to deserialize a bearer token from an HTTP request generically.
- User session fetch/user identity
- Login/logout routing through OAuth
- Bringing app-specific storage layer back online through the newly-updated extension protocols for the backend.

Here we are testing the behavior of the backend Node.js service when the derived app does not specify any extension inputs for user identity or session management (i.e. they take the defaults). This should just work obviously. But required drilling down through holism which is very painful.

Next step, is to start verifying the correct application of app-defined filter specs to the holism filters that box app-defined user ID and session callbacks.

Then, verify the correct integration of the user ID and session callbacks (has to be done simultanesouly w/re-integrating app-specific backend subsystems through what we call the 'appStateContext' object. Someday this will all be cells...