WIP v0.0.50-crystallite buildID "sf3yCXlfRQOG-PY3o4zkVQ"
DO NOT ATTEMPT TO USE THIS WIP TEST RELEASE PLEASE!

This commit is a baseline an updated baseline for v0.0.50-crystallite
that includes many interlocking and related fixes to HolisticServiceCore, HolisticNodeService, and HolisticHTML5Service intended to stabilize and deal with the matrix of possibible initial conditions that have to be sorted in order to determine how/if to start hashroute processing in the HolisticHTML5Service kernel and derived app service.

It is expected that if you work w/this specific test build that you will not see any calls to into your derived app's hashroute lifecycle action from the HolisticHTML5Service_Kernel process because it is disabled in this build.

Subsequent release (later today) will restore the hashroute lifecycle action dispatch (but it will be triggered by modified logic).