TEST RELEASE v0.0.50-crystalliate buildID "xT5lKFfNQuyil9ynScGPCA"
This work is "code complete" and I think it's a good design and working okay. However, the new classes will need to be bulletproof. I need help getting test coverage on @encapsule/holarchy updates here specifically.

New exports:

- cmasHolarchyPackage - An instance of CellModelArtifactSpace specialized for @encapsule/holarchy RTL package.
- CellModelArtifactSpace - An ES6 class instantiated to create a "CellModel artifact space" which is defined as the set of all possible unique CellModel artifact IRUT ID's that may be induced from a set of developer-defined "CellModel label" strings.
- CellModelTemplate - An ES6 class instantiated to create a "CellModel" generator for synthesizing specific members of a specific "family" of uniquely-specialized CellModel(s) that are related insofar as they are
    - Synthesized with specific, developer-defined CellModel generator algorithm
    - Follow consistent conventions wrt CellModelArtifactSpace (note that CellModelTemplate extends CellModelArtifactSpace).

This work is "in-progress" - it's the first application of the new CellModelArtifactSpace and CellModelTemplate ES6 classes. At this point, it is high-level structural.

The actual implementation of the cmtObservableValue generator IS NOT DONE.
The actual implementation of the cmtObservableValueWorkerProxy generator IS NOT DONE.

I am actively working on these today to get them to 100% (at which point these artifact "families") can be used in derived code (and inside of new CellModelTemplate's that are defined to deal w/application-service-specific feature work).

New exports:

- cmasHolarchyCMPackage - An instance of CellModelArtifactSpace specialized for @encapsule/holarchy-cm RTL package.
- cmtObservableValue - An instance of CellModelTemplate specialized for synthesizing members of the family of CellModel "ObservableValue"
- ObservableValueProxy - A CellModel that represents an arbitrary proxy to any member of the "ObservableValue" family of CellModel defined by cmtObservableValue

Ignore this for now... In small detail, I have converted previous experiment done in HolisticHTML5Service_DOMLocation w/an "ObservableValue" CellModel helper (composed into DOMLocation's cell memory via its APM) to use the new infrastructure above

HolisticHTML5Service_DOMLocation CellModel now includes an ObservableValue specialized for RouterEventDescriptor value type. It's not connected to anything yet (and the underlying synthesized CellModel is incomplete as above).

This will be evolving in to clearer focus in next wave of updates (@encapsule/holistic vNext (v0.0.51)).