TEST RELEASE v0.0.52-tourmaline buildID "3Qlaxx2-Tp2sgq_INe74NA"
WIP on @encapsule/holarchy-cm

This commit includes support for the following scenario:

Configure an ObservableValueHelper cell instance either by providing the configuration data when the ObservableValueHelper cell is activated. Or, via its configure action.

After the ObservableValueHelper cell instance is "linked":

We now support operator ObservableValueHelper isValueActive (or we will very shortly after this commit).

OVH isValueActive returns false iff OVH step is not "linked"

Provided "linked", then we delegate to our ObservableValueWorker cell process instance the same question; isValueActive via its same-name operator.

Because we're cautious, OVW isValueActive returns false iff OVW step is not "proxy-connected" (similar to policy in the layer above it).

Provided "proxy-connected", then we're quite certain we can communicate w/the provider cell process via our ovcpProviderProxy helper cell.

So, we delegate through our proxy to the provider cell process a new operator request provided by ObservableValueBase (but intended to be used by a provider) that actually (will doesn't yet but it's called) check OCD at the indicated coordinates to determine if the ObservableValue ACTUALLY EXISTS or not.

This seems like a lot of work. And, it is. But, sharing memory w/out writing code and tests is worth...

Query of ObservableValueHelper valueIsActive operator