TEST RELEASE v0.0.53-malachite buildID "MhFm8tPBSWaHgQcGkRUWcg"
Update to @encapsule/holarchy-cm ObservableValueBase:

- remove constrain that value type spec passed to ObservableValue_T::synthesizeCellModel must be default activatable (i.e. if you pass nothing to a filter that uses your value type spec as inputFilterSpec will it fail?) Previously, if the answer is YES then ObervableValueBase reset fails. If not, then okay.

Now, we box the actual type-specialized value in a default-constructable parent object allowing us to default construct the OV cell instance on reset w/out regard to the value type spec.

And, we kept revision out of `mailbox` to limit impact.

And, if the value has never been written, then mailbox.value === undefined which is what we want! Unless of course the developer has specified a value type spec that provides a default value if passed undefined. And, then we'll get that in `mailbox.value` on reset.

Note: THIS IS EXACTLY WHAT WE WANT EXPECT; it will not have been written and it's revision count will be -1 (in default state).

This is like this specifically for React integration use cases BTW...